+++ src/app/models/virus-interactions.ts
@@ -1,9 +1,30 @@
+import {PharosSerializer} from "./pharos-base";
 
 export class ViralInteractionDetails {
   finalLR: number;
   protein_name: string;
   protein_ncbi: string;
   dataSource: string;
+  pdbIDs: string[] = [];
+  confirmed(): boolean{
+    return this.pdbIDs?.length > 0;
+  }
+  static sort(a: ViralInteractionDetails, b: ViralInteractionDetails) : number {
+    if(a.confirmed() && !b.confirmed()){
+      return -1;
+    }
+    if(b.confirmed() && !a.confirmed()){
+      return 1;
+    }
+    return b.finalLR - a.finalLR;
+  }
+  constructor(obj: ViralInteractionDetails = {} as ViralInteractionDetails) {
+    this.finalLR = obj.finalLR;
+    this.protein_name = obj.protein_name;
+    this.protein_ncbi = obj.protein_ncbi;
+    this.dataSource = obj.dataSource;
+    this.pdbIDs = obj.pdbIDs;
+  }
 }
 
 export class VirusDetails {
@@ -16,5 +37,36 @@ export class VirusDetails {
   subfamily: string;
   genus: string;
   species: string;
-  interactionDetails: [ViralInteractionDetails];
+  interactionDetails: ViralInteractionDetails[] = [];
+  confirmed(): boolean{
+    return !!this.interactionDetails.find(ppi => {return ppi.confirmed()});
+  }
+  static sort(a: VirusDetails, b: VirusDetails) : number {
+    if(a.confirmed() && !b.confirmed()){
+      return 1;
+    }
+    if(b.confirmed() && !a.confirmed()){
+      return -1;
+    }
+    return a.interactionDetails.length - b.interactionDetails.length;
+  }
+}
+
+export class VirusDetailsSerializer implements PharosSerializer{
+  fromJson(json: any): any {
+    const obj = new VirusDetails();
+    Object.entries((json)).forEach((prop) => obj[prop[0]] = prop[1]);
+    const viralInteractionDetailsSerializer = new VirusInteractionSerializer();
+    obj.interactionDetails = json.interactionDetails.map(ppi => {return viralInteractionDetailsSerializer.fromJson(ppi)})
+      .sort((a,b) => {return ViralInteractionDetails.sort(a,b);});
+    return obj;
+  }
+}
+
+export class VirusInteractionSerializer implements PharosSerializer{
+  fromJson(json: any): any {
+    const obj = new ViralInteractionDetails();
+    Object.entries((json)).forEach((prop) => obj[prop[0]] = prop[1]);
+    return obj;
+  }
 }