+++ src/app/pharos-main/data-details/target-details/panels/level-summary-panel/levels/tclin-summary/tclin-summary.component.ts
@@ -1,5 +1,6 @@
 import {ChangeDetectionStrategy, Component, Input} from '@angular/core';
 import {Target} from '../../../../../../../models/target';
+import {DynamicPanelBaseComponent} from "../../../../../../../tools/dynamic-panel-base/dynamic-panel-base.component";
 
 /**
  * shows details about tclin targets
@@ -10,7 +11,7 @@ import {Target} from '../../../../../../../models/target';
   styleUrls: ['./tclin-summary.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class TclinSummaryComponent {
+export class TclinSummaryComponent extends DynamicPanelBaseComponent{
   /**
    * input target
    */
@@ -21,23 +22,10 @@ export class TclinSummaryComponent {
    */
   @Input() data: any;
 
-  @Input() apiSources: any[];
-
   /**
    * no args constructor
    */
   constructor() {
-
-  }
-
-  getTooltip(label: string): string {
-    if (this.apiSources) {
-      const tooltip = this.apiSources.filter(source => source.field === label);
-      if (tooltip.length) {
-        return tooltip[0].description;
-      } else {
-        return null;
-      }
-    }
+    super();
   }
 }