+++ src/app/pharos-main/data-list/cards/target-card/target-card.component.ts
@@ -1,6 +1,7 @@
 import {ChangeDetectionStrategy, Component, Input, OnInit} from '@angular/core';
 import {Target} from '../../../../models/target';
 import {DataProperty} from "../../../../tools/generic-table/components/property-display/data-property";
+import {DynamicPanelBaseComponent} from "../../../../tools/dynamic-panel-base/dynamic-panel-base.component";
 
 /**
  * component to display a consolidated target view
@@ -11,19 +12,21 @@ import {DataProperty} from "../../../../tools/generic-table/components/property-
   styleUrls: ['./target-card.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
-export class TargetCardComponent implements OnInit {
+export class TargetCardComponent extends DynamicPanelBaseComponent implements OnInit {
   /**
    * target to display. optional because it may also get injected
    */
   @Input() target?: Target;
-  @Input() apiSources: any[];
+
   /**
    * boolean to display knowedge chart under radar chart
    * @type {boolean}
    */
   @Input() showKnowledge = false;
 
-  constructor() {}
+  constructor() {
+    super();
+  }
 
   /**
    * if target exists, fetch radar chart data
@@ -41,16 +44,5 @@ export class TargetCardComponent implements OnInit {
   getDisplayProp(prop: DataProperty){
     prop.tooltip = this.getTooltip(prop.label);
     return prop
-  }
-
-  getTooltip(label: string): string {
-    if (this.apiSources) {
-      const tooltip = this.apiSources.filter(source => source.field === label);
-      if (tooltip.length) {
-        return tooltip[0].description;
-      } else {
-        return null;
-      }
-    }
   }
 }