+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# 3.2.0 (2020-11-09)
+### Added Features
+* ** Pathways for target details pages
+    * Show Reactome Pathway browser when possible
+    * Show other pathway datasources and links to their pages
+    * Link to targets lists of all targets in each pathway
+* ** Show PDB ids for virus interactions that have confirmed PDB ids
+    * Allows distinction between predicted viral interactions, and confirmed viral interactions
+    
+### Bug Fixes
+* fixes for displaying on small screens
+* accept queries from /search?q=term links
+
+
 # 3.1.0 (2020-10-02)
 ### Added Features
 * **Better data provenance:**
@@ -45,7 +59,7 @@ _______________
 ### Added Features
 * **IDG Resources:**
     * Fetching and Displaying IDG resources for Cell Lines, Constructs, Mouse Lines, Datasets, etc. from <a href="https://rss.ccs.miami.edu/rss-apis/" target="_blank">https://rss.ccs.miami.edu/rss-apis/</a>
-    * Display expression with a mouse anatamogram
+    * Display expression with a mouse anatomogram
 
 * **Predicted Viral Interactions:**
     * Showing Target List Facets, and Target Details Panel for Predicted Viral Interactions from <a href="http://phipster.org/" target="_blank">P-HIPSTer</a>.